moved client implementation into own package to ensure its only used indirectly
